# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReadinessStatusStatus =
  (typeof ReadinessStatusStatus)[keyof typeof ReadinessStatusStatus];

export const ReadinessStatusStatus = {
  ready: "ready",
  not_ready: "not_ready",
} as const;

export interface ReadinessStatus {
  status: ReadinessStatusStatus;
}

export type ApiPayload =
  { [key: string]: unknown } | { [key: string]: unknown }[];

export type AuthMeResponseRole =
  (typeof AuthMeResponseRole)[keyof typeof AuthMeResponseRole];

export const AuthMeResponseRole = {
  PATIENT: "PATIENT",
  CLINICIAN: "CLINICIAN",
  ADMIN: "ADMIN",
} as const;

export interface AuthMeResponse {
  userId: string;
  role: AuthMeResponseRole;
  /** @nullable */
  organizationId?: string | null;
  capabilities: string[];
}

export type AccessDecisionDecision =
  (typeof AccessDecisionDecision)[keyof typeof AccessDecisionDecision];

export const AccessDecisionDecision = {
  approved: "approved",
  denied: "denied",
} as const;

export interface AccessDecision {
  decision: AccessDecisionDecision;
}

export type QrResolveScopesItem =
  (typeof QrResolveScopesItem)[keyof typeof QrResolveScopesItem];

export const QrResolveScopesItem = {
  visits: "visits",
  prescriptions: "prescriptions",
  labs: "labs",
} as const;

export interface QrResolve {
  /** @maxLength 128 */
  token: string;
  /**
   * @minItems 1
   * @maxItems 3
   */
  scopes?: QrResolveScopesItem[];
  /**
   * @minimum 5
   * @maximum 240
   */
  durationMinutes?: number;
}

export interface EncounterInput {
  /** @pattern ^[a-fA-F0-9]{24}$ */
  patientId: string;
  /** @maxLength 200 */
  diagnosis: string;
  /** @maxLength 500 */
  reason: string;
  date: string;
}

export interface PrescriptionInput {
  /** @pattern ^[a-fA-F0-9]{24}$ */
  patientId: string;
  /** @maxLength 200 */
  drug: string;
  /** @maxLength 120 */
  dose: string;
  /** @maxLength 120 */
  frequency: string;
  start: string;
  end: string;
}

export interface ObservationInput {
  /** @pattern ^[a-fA-F0-9]{24}$ */
  patientId: string;
  /** @maxLength 200 */
  title: string;
  /** @maxLength 200 */
  value: string;
  /** @maxLength 80 */
  unit: string;
  date: string;
}

export type ErrorResponseError = {
  code: string;
  message: string;
  requestId?: string;
};

export interface ErrorResponse {
  error: ErrorResponseError;
}

/**
 * Successful API response
 */
export type ApiSuccessResponse = ApiPayload;

/**
 * Missing or invalid Clerk session
 */
export type UnauthenticatedResponse = ErrorResponse;

/**
 * Role, organization, or consent denied
 */
export type ForbiddenResponse = ErrorResponse;

/**
 * Request validation failed
 */
export type BadRequestResponse = ErrorResponse;

export type GetAdminTrendsParams = {
  condition?: string;
  /**
   * @pattern ^\d{4}-(0[1-9]|1[0-2])$
   */
  startMonth?: string;
  /**
   * @pattern ^\d{4}-(0[1-9]|1[0-2])$
   */
  endMonth?: string;
};