# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get K-suppressed aggregate system summary
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/trends:
    get:
      operationId: getAdminTrends
      tags: [admin]
      summary: Query privacy-protected disease trends
      security:
        - bearerAuth: []
      parameters:
        - name: condition
          in: query
          schema: { type: string }
        - name: startMonth
          in: query
          schema: { type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])$" }
        - name: endMonth
          in: query
          schema: { type: string, pattern: "^\\d{4}-(0[1-9]|1[0-2])$" }
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/regions:
    get:
      operationId: getAdminRegions
      tags: [admin]
      summary: Get privacy-protected regional aggregates
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/conditions:
    get:
      operationId: getAdminConditions
      tags: [admin]
      summary: Get condition category aggregates
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/privacy-config:
    get:
      operationId: getAdminPrivacyConfig
      tags: [admin]
      summary: Get privacy threshold and aggregate metrics
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/audit:
    get:
      operationId: getAdminAuditEvents
      tags: [admin]
      summary: Get paginated metadata-only audit events
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/security-events:
    get:
      operationId: getAdminSecurityEvents
      tags: [admin]
      summary: Get safe security event summaries
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ApiSuccess:
      description: Successful API response
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiPayload" }
...
    Forbidden:
      description: Role, organization, or consent denied
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ApiPayload:
      oneOf:
        - type: object
          additionalProperties: true
        - type: array
          items:
            type: object
            additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
          required: [code, message]
      required: [error]